requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sentence-transformers==2.2.2
faiss-cpu==1.8.0
python-multipart==0.0.6
pymupdf==1.22.5
tqdm==4.66.1
rank_bm25==0.2.2
scikit-learn==1.3.2
pandas==2.2.2